prep for REL_1_1_1_Nov_26_2013 release

Conflicts:
	CMakeLists.txt
	configure.in
	src/PThreads.cpp
	src/RemoteSyslogAppender.cpp

+++ src/PThreads.cpp
@@ -6,13 +6,13 @@ namespace log4cpp {
     namespace threading {
 
         char* getThreadId(char* buffer) {
-            ::sprintf(buffer, "%ld", (long int)pthread_self());
+            ::sprintf(buffer, "%lu", pthread_self());	// thread id unsigned
             return buffer;
         }
 
         std::string getThreadId() {
             char buffer[16];
-            ::sprintf(buffer, "%ld", (long int)pthread_self());
+            ::sprintf(buffer, "%lu", pthread_self());	// thread id unsigned
             return std::string(buffer);     
         }
